error handling and authorization

+++ backend/routes/posts.js
@@ -37,7 +37,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(result => {
       res.status(201).json({
@@ -47,6 +48,11 @@ router.post(
           id: result._id
         }
       });
+    })
+    .catch(err => {
+      res.status(500).json({
+        message: 'Creating a post failed'
+      })
     });
 });
 
@@ -63,12 +69,24 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     })
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-      res.status(200).json({
-        message: 'Update successful'
-      });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if (result.nModified) {
+        res.status(200).json({
+          message: 'Update successful'
+        });
+      } else {
+        res.status(401).json({
+          message: 'Not authorized'
+        });
+      }
+    })
+    .catch(err => {
+      res.status(500).json({
+        message: 'Could not update post'
+      })
     });
 });
 
@@ -92,6 +110,11 @@ router.get("", (req, res, next) => {
       posts: fetchedPosts,
       total: count
     });
+  })
+  .catch(err => {
+    res.status(501).json({
+      message: "Fetching posts failed"
+    });
   });
 });
 
@@ -105,12 +128,26 @@ router.get("/:id", (req, res, next) => {
       })
     }
   })
+  .catch(err => {
+    res.status(500).json({
+      message: "Fetching post failed"
+    });
+  });
 })
 
 router.delete("/:id", checkAuth, (req, res) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({message:"Deleted"});
-  });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n) {
+      res.status(200).json({message:"Deleted"});
+    } else {
+      res.status(401).json({message:"Not authorized"});
+    }
+  })
+  .catch(err => {
+    res.status(501).json({
+      message: "Deleting post failed"
+    });
+  });;
 });
 
 module.exports = router;
